backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
dnspython==2.4.2
pydantic==2.5.2
